requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.1
django-cors-headers==4.7.0
django-heroku==0.3.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
